Move package config to init.el and configure Org mode
Use Org mode as provided by the Orgmode.org
ELPA (org-plus-contrib). In order to correctly load this Org mode
instead of the one shipped with Emacs, the package system needs to be
initialized before executing any Org functions. Therefore the entire
setup of the package system has to be moved to init.el.

Some initial setup of agenda files and the like is also contained in
this commit.